msm-nand: Add dma_sync_single* calls in msm_nand_read_oob.
When empty pages are read, certain bytes in the buffer that is read are set
0x54 instead of 0xFF. This is due to the ECC error correction logic of the
nand controller. CPU needs to correct these bytes back to 0xFF and it needs
to access these bytes while the buffer is still owned by the device. This
can be achieved if dma_sync_single_for_cpu and dma_sync_single_for_device
are called before & after the CPU access to the buffer owned by the device.

Signed-off-by: Brent DeGraaf <bdegraaf@quicinc.com>