msm: acpuclock: Fix a bug in overclocking code.
Note: This effectively reverts:
7321b17c761c1c3d6ef7ed9bb6de927345274761

History:
This was done, due to the mass errors from the inability to reduce the AXI
rate, as it was shipped from AOSP in the early part of .35
(it was re-tested a few months back and produced the same).

With Cyanogen's AXI scaling port and stepping freq. fixup, this may not
be an issue.

Re-evaluating...